Add expected ISO8601 dates to tests
If there is no year in the input date, the ISO8601 compliant parsed
result should indicate XXXX for the year. Even though a scholar might
be able to infer the actual year from context.

Adding this column gives a better overview of what the parser
correctly parses (but which might still be incorrect once adjusted
for context).